Move ContextServiceMock to services package
